Filling out README.md
It's mostly there, but need a few links and a finalized API for the UDF Creator. We can update that easily later though if needed.

Trying out github website editing of this markdown -- I'm not sure I'm committing to the right branch (should it be in my fork?) but oh well... hopefully it works.