ML-3109 - added vgg16 and vgg19 to DeepImageFeaturizer (#76)
Added vgg16 and vgg19 to DeepImageFeaturizer. Note: Had to disable VGG16 and VGG19 from travis as they run into OOM issue there and memory in travis can not be changed. For running the tests locally, it is recommended that one runs all tests using the environment variable (see code in PR).