[ML-3035] Add persistence & test for DeepImageFeaturizer in Scala (#86)
Adds persistence for Scala implementation of DeepImageFeaturizer. Since DeepImageFeaturizer stores all its data as instances of org.apache.spark.ml.Param, all we need to do is extend Spark's default ML persistence traits (DefaultParamsReadable, DefaultParamsWritable).

See MLlib's Binarizer and BinarizerSuite for an example of a similar Transformer/test suite pair.

As is, this PR enables DeepImageFeaturizer persistence in Scala but not Python; to enable it in Python, we just need to inherit from JavaMLWriter & JavaMLReader, which define Python persistence methods (save, load) that call the corresponding Scala methods. However, we need to ensure the behavior in python and scala are consistent before we support persistence read/write in python.